Secure the application

+++ app/api/issues/[id]/route.ts
@@ -1,51 +1,57 @@
 import { issueSchema } from "@/app/validationSchemas";
 import prisma from "@/prisma/client";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
-
 
 export async function PATCH(
   request: NextRequest,
-  { params } : { params : { id : string }}) {
+  { params }: { params: { id: string } }
+) {
+  const session = await getServerSession();
+  if (!session) return NextResponse.json({}, { status: 401 });
+
   const body = await request.json();
   const validation = issueSchema.safeParse(body);
 
-  if (!validation.success){
-    return NextResponse.json(validation.error.format(), { status : 400 });
+  if (!validation.success) {
+    return NextResponse.json(validation.error.format(), { status: 400 });
   }
 
   const issue = await prisma.issue.findUnique({
-    where: { id : parseInt(params.id)}
+    where: { id: parseInt(params.id) },
   });
 
-  if(!issue)
-    return NextResponse.json({ error: 'Invalid issue' }, {status: 404});
+  if (!issue)
+    return NextResponse.json({ error: "Invalid issue" }, { status: 404 });
 
   const updatedIssue = await prisma.issue.update({
-    where: {id : issue.id},
+    where: { id: issue.id },
     data: {
       title: body.title,
-      description: body.description
-    }
+      description: body.description,
+    },
   });
 
   return NextResponse.json(updatedIssue);
-}  
+}
 
 export async function DELETE(
   request: NextRequest,
-  { params } : { params : { id : string }}) {
+  { params }: { params: { id: string } }
+) {
+  const session = await getServerSession();
+  if (!session) return NextResponse.json({}, { status: 401 });
 
-    const issue = await prisma.issue.findUnique({
-      where: { id: parseInt(params.id) }
-    });
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(params.id) },
+  });
 
-    if(!issue)
-      return NextResponse.json({ error: 'Invalid Issue' }, { status: 404 } );
+  if (!issue)
+    return NextResponse.json({ error: "Invalid Issue" }, { status: 404 });
 
-    await prisma.issue.delete({
-      where: {id: issue.id}
-    });
+  await prisma.issue.delete({
+    where: { id: issue.id },
+  });
 
-    return NextResponse.json({});
-
-  }
+  return NextResponse.json({});
+}
